refactor(ai): restore existing Responses code locations
